% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_tools.R
\name{extract_date}
\alias{extract_date}
\title{Extract dates from text}
\usage{
extract_date(v)
}
\arguments{
\item{v}{Text vector}
}
\value{
A list of dates
}
\description{
Sometimes dates can be contained in text,
this function extracts those dates from text.
Please also see `messydates::as_messydate()` for more on dates extraction
from texts.
}
\examples{
extract_date("This function was created on the 29 September 2021")
}
