% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_tools.R
\name{extract_locations}
\alias{extract_locations}
\title{Extract locations from strings}
\usage{
extract_locations(v)
}
\arguments{
\item{v}{Text vector.}
}
\value{
A data frame of locations and the number of times they appear.
}
\description{
Extract locations from strings
}
\details{
The function relies on geographical entity detection from NLP models.
}
\examples{
#extract_locations(c("This is the United States", "This is Sao Paulo",
#"I was in Rio de Janeiro and Sao Paulo, then back to the United States"))
}
