% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_tools.R
\name{extract_names}
\alias{extract_names}
\title{Extract a list of possible names of individuals in texts}
\usage{
extract_names(v)
}
\arguments{
\item{v}{A text vector.}
}
\value{
A data frame of individual names and the number of times they appear.
}
\description{
Extract a list of possible names of individuals in texts
}
\details{
The function relies on named entity recognition from NLP models.
}
\examples{
#extract_names("This package was created by Jael, James, and I.")
}
