% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment.R
\name{extract_promises}
\alias{extract_promises}
\title{Extract future promises from political discourses}
\usage{
extract_promises(v)
}
\arguments{
\item{v}{Text vector.}
}
\description{
Extract future promises from political discourses
}
