% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment.R
\name{extract_related_terms}
\alias{extract_related_terms}
\title{Extract terms related to subjects}
\usage{
extract_related_terms(v, subjects, n = 5)
}
\arguments{
\item{v}{Text vector.}

\item{subjects}{List of subjects.}

\item{n}{Number of terms.}
}
\description{
Extract terms related to subjects
}
