% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment.R
\name{extract_similarities}
\alias{extract_similarities}
\title{Extract similarities and differences in texts/segments}
\usage{
extract_similarities(v, comparison, method)
}
\arguments{
\item{v}{Text vector}

\item{comparison}{Would you like to extract similarities or differences
between treaties?
If not specified, defaults to "similarities".
Alternatively, users can also get the "differences".}

\item{method}{A method for checking similarities or differences.
If chosen comparison are similarities, extracts "correlation" between
treaty texts if not specified.
Other similarity methods from `quanteda.textstats::textstat_simil()`
include "cosine", "jaccard", "ejaccard", "dice", "edice",
"simple matching", and "hamann".
If chosen comparison are differences, extracts "euclidean" difference
between treaty texts if not specified.
Other difference methods from `quanteda.textstats::textstat_dist()` include
"manhattan", "maximum", "canberra", and "minkowski".}
}
\description{
Extract similarities and differences in texts/segments
}
