% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_tools.R
\name{extract_speaker}
\alias{extract_speaker}
\title{Extract a list of the speakers in texts}
\usage{
extract_speaker(v)
}
\arguments{
\item{v}{A text vector.}
}
\value{
A list of speakers.
}
\description{
Extract a list of the speakers in texts
}
\details{
The function relies on NLP models and, therefore, results
might not be accurate or consistent.
}
\examples{
\dontrun{
extract_speaker(US_News_Conferences_1960_1980[20, 3])
}
}
