% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_tools.R
\name{extract_split}
\alias{extract_split}
\title{Extract splited texts}
\usage{
extract_split(text, splitsign = "\\\\.")
}
\arguments{
\item{text}{text variable}

\item{splitsign}{Where do you want to split?
By default sentences (".").
This can also be words, signals or other markers you want.
For special characters, please use escape sign before (i.e. "\\").}
}
\value{
A splitted list for each row
}
\description{
Split texts into structured lists according to a split sign.
}
\examples{
text <- "This is the first sentence. This is the second sentence."
extract_split(text)
}
