% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_tools.R
\name{extract_title}
\alias{extract_title}
\title{Extract title from text}
\usage{
extract_title(v)
}
\arguments{
\item{v}{Text variable/object}
}
\value{
A list of the first sentences
}
\description{
A lot of information is contained in the first sentence of a text.
For political texts, dates and locations are often contained in the
first sentence.
This function extracts the first sentence of text.
}
\examples{
text <- "This is the first sentence. This is the second sentence."
extract_title(text)
}
