% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic.R
\name{extract_topics}
\alias{extract_topics}
\title{Extract topic from political discourses}
\usage{
extract_topics(.data, dictionary = "CAP")
}
\arguments{
\item{.data}{A data frame, promises, or text vector.
For data frames, function will search for "text" variable.
For promises data, function will search for "promises" variable.}

\item{dictionary}{The dictionary of 20 major political topics from the
Comparative Agendas Project is used by default.
Users can also declare a custom dictionary as a vector or a list.
If users declare a vector, each element is treated as a independent topic.
If users declare a list of subjects and related terms, function understands
names as topic and words as terms.}
}
\description{
Extract topic from political discourses
}
\examples{
\donttest{
extract_topics(US_News_Conferences_1960_1980[1:5, 3])
extract_topics(US_News_Conferences_1960_1980[1:5, 3],
               dictionary = c("military", "development"))
extract_topics(.data = US_News_Conferences_1960_1980[1:5, 3],
               dictionary = list("military" = c("military", "gun", "war"),
                                 "development" = c("development", "interest rate", "banks")))
}

}
