% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic.R
\name{gather_topics}
\alias{gather_topics}
\title{Gather topic from political discourses}
\usage{
gather_topics(v, dictionary = "CAP")
}
\arguments{
\item{v}{Text vector or annotated data frame.
If missing, opens the political topics codebook.}

\item{dictionary}{The dictionary of 20 major political topics from the
Comparative Agendas Project (Jones et al., 2023) is used by default.
Users can also declare a custom dictionary as a vector or a list.
If users declare a vector, each element is treated as a independent topic.
If users declare a list of subjects and related terms, function understands
names as topic and words as terms.
For more information on how the CAP topics were adapted, please run
`gather_topics()` to access the political topics codebook.}
}
\value{
A list of topics present in each text separated by comma.
}
\description{
Gather topic from political discourses
}
\examples{
\donttest{
summary(gather_topics(US_inaugural_addresses_1993_2025$text))
}
}
