% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_speaker.R
\name{get_speakers}
\alias{get_speakers}
\title{Split text by speakers}
\source{
https://stackoverflow.com/questions/41100482/split-speaker-and-dialogue-in-rstudio
The above was used as a reference.
}
\usage{
get_speakers(textvec, splitsign = "\\\\.", speakermark = ":")
}
\arguments{
\item{textvec}{text vector}

\item{splitsign}{sign or string used to split text vector.
Line markers ("/n") by default.
Often these are line spaces but depending on text different signs
or even string can be chosen.}

\item{speakermark}{speaker marker sign.
Colon (":") by default.
Often these are colon but depending on text different
or even string can be chosen (i.e. "Thank you").}
}
\value{
All text aggregated by speaker.
}
\description{
Split text by speakers
}
\details{
This is an interactive function.
It asks users which speakers are legit, or not.
For legit speakers, text is aggregated altogether in one vector.
}
