% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urgency.R
\name{get_urgency}
\alias{get_urgency}
\title{Urgency Analysis}
\usage{
get_urgency(v, summarise = "sum")
}
\arguments{
\item{v}{Text vector or annotated data frame.}

\item{summarise}{How to handle multiple matches for the same dictionary
in the same text observation?
By default, multiple matches are added together and
their "sum" per text observation is returned.
Users can, instead, choose the "mean", which returns the average
score per dictionary per text observation, the "max" or the "min",
when there are multiple matches.}
}
\value{
A scored data frame for each dimension of urgency.
}
\description{
Urgency Analysis
}
\details{
Urgency in political discourses is an expression of how necessary and/or
how soon an action should be undertaken or completed.
This is measured along four dictionaries,
two related to necessity (e.g., degree of intensity and of commitment)
and two related to timing (e.g., frequency and timing of action).
The dictionaries include terms for each of these dimensions of urgency.
Their scores were developed through an online survey that took place
in June 2026. The survey results were recorded as counts of the number of
participants who selected an urgency-related word as more urgent than its pair.
To analyze the survey results, we employed Bradley-Terry models for
paired comparisons. We use raw log odds coefficients to calculate urgency.
Urgency scores for each disctionary are returned.
The aggregated urgency scores is calculated by adding these scores and
then divided by the coefficient for the word "must", to ensure
interpretability (i.e., an aggregated score of 1 in equivalent to "we must do").
Users are welcome to change how urgency scores are aggregated and transform these
scores into probabilities or odds ratios as necessary.
}
\examples{
\donttest{
get_urgency(US_inaugural_addresses_1993_2025$text)
#a = get_urgency(select_priorities(US_inaugural_addresses_1993_2025$text))
#summary(get_urgency(US_inaugural_addresses_1993_2025$text))
#plot(get_urgency(US_inaugural_addresses_1993_2025$text))
}
}
