% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urgency2.R
\name{get_urgency2}
\alias{get_urgency2}
\title{Alternative urgency function}
\usage{
get_urgency2(.data)
}
\arguments{
\item{.data}{A data frame, priorities data frame coded using
`select_priorities()`, or text vector.
For data frames, function will search for "text" variable.
For priorities data frame function will search for "priorities" variable.
If missing, opens the webpage containing the urgency codebook.}
}
\description{
Alternative urgency function
}
\examples{
get_urgency2(US_News_Conferences_1960_1980[1:10, 3])
}
