% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urgency.R
\name{get_urgency_rank}
\alias{get_urgency_rank}
\title{Rank urgent topics}
\usage{
get_urgency_rank(v, subjects)
}
\arguments{
\item{v}{Text vector or annotated data frame.}

\item{subjects}{List of subjects.}
}
\description{
Rank urgent topics
}
\examples{
\donttest{
get_urgency_rank(US_News_Conferences_1960_1980[1:10,3])
}
}
