% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_tools.R
\name{load_pdf}
\alias{load_pdf}
\title{Load text from PDFs}
\usage{
load_pdf(path)
}
\arguments{
\item{path}{The path to a PDF file or a folder containing multiple PDFs.}
}
\value{
A list of texts.
}
\description{
Load text from PDFs
}
