% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urgency.R
\name{rank_urgent_topics}
\alias{rank_urgent_topics}
\title{Rank urgent topics}
\usage{
rank_urgent_topics(v, subjects)
}
\arguments{
\item{v}{Text vector or annotated data frame.}

\item{subjects}{List of subjects.}
}
\description{
Rank urgent topics
}
\examples{
\donttest{
rank_urgent_topics(US_News_Conferences_1960_1980[1:10,3])
}
}
