% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_tools.R
\name{re_order}
\alias{re_order}
\title{Helper function for re-ordering dates}
\usage{
re_order(l)
}
\arguments{
\item{l}{List of dates}
}
\value{
A list of dates of the same lenghte but re-ordered, if necessary.
}
\description{
Helper function for re-ordering character dates for consistency.
Default is DMY.
}
