% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_tools.R
\name{read_pdf}
\alias{read_pdf}
\title{Read text from PDFs}
\usage{
read_pdf(path)
}
\arguments{
\item{path}{The path to a PDF file or a folder containing multiple PDFs.}
}
\value{
A list of texts.
}
\description{
Read text from PDFs
}
