% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priorities.R
\name{select_priorities}
\alias{select_priorities}
\title{Select future priorities from political discourses}
\usage{
select_priorities(v, na.rm = TRUE)
}
\arguments{
\item{v}{Text vector or annotated data frame.}

\item{na.rm}{Would you like political statements that do not contain a
political action to be removed?
By default, TRUE.}
}
\value{
A data frame with syntax information by sentences and
a variable identifying which of these sentences are priorities.
}
\description{
Political priorities are statements in which actors express their
intent or commitment to take political action in the future.
}
\examples{
#select_priorities(US_inaugural_addresses_1993_2025$text)
}
