% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promises.R
\name{select_promises}
\alias{select_promises}
\title{Select future promises from political discourses}
\usage{
select_promises(.data)
}
\arguments{
\item{.data}{A (annotated) data frame or text vector.
For data frames, function will search for "text" variable.
For annotated data frames, please declare an annotated data frame
at the sentence level.}
}
\value{
A data frame with syntax information by sentences and
a variable identifying which of these sentences are promises.
}
\description{
Political promises are statements in which actors express their
intent or commitment to take political action in the future.
}
\examples{
#select_promises(US_News_Conferences_1960_1980[1:2,3])
}
