% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_urgency.R
\name{sim_urgency}
\alias{sim_urgency}
\title{Simulating urgency}
\usage{
sim_urgency(urgency, commitment, intensity, timing, frequency, pronoun = "We")
}
\description{
Simulating urgency
}
\examples{
sim_urgency()
sim_urgency(urgency = 0.5)
sim_urgency(urgency = 2.5)
sim_urgency(urgency = -2.5)
sim_urgency(commitment = 0.6)
sim_urgency(commitment = 0.6, intensity = 1.4)
sim_urgency(commitment = 0.6, intensity = 1.4, timing = 1.4)
}
