% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_tools.R
\name{split_text}
\alias{split_text}
\title{Split texts}
\usage{
split_text(v, splitsign = "\\\\.")
}
\arguments{
\item{v}{Text vector or annotated data frame.}

\item{splitsign}{Where do you want to split?
By default sentences (".").
This can also be words, signals or other markers you want.
For special characters, please use escape sign before (i.e. "\\").}
}
\value{
A list of lists the same length as vector.
}
\description{
Split texts into structured lists of lists according to a split sign.
}
\examples{
\donttest{
split_text("This is the first sentence. This is the second sentence.")
}
}
