% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_tools.R
\name{text_match}
\alias{text_match}
\title{Get text matches}
\usage{
text_match(text, match, invert = FALSE, ignore.case = TRUE)
}
\arguments{
\item{text}{A text variable}

\item{match}{A regex match for a word(s) or expression.
For multiple words, please use "|" to divide them.}

\item{invert}{Do you want texts without certain matches to be returned?
By default FALSE.}

\item{ignore.case}{Shoud case be ignored?
By default, TRUE.}
}
\value{
A list of matches of the same length as text variable
}
\description{
Get texts in which only certain "matches" occur.
}
\examples{
text <- c("This function was created on the 29 September 2021",
"Today is October 12, 2021")
text_match(text, "October")
}
