% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/US_inaugural_addresses_1993_2025.R
\docType{data}
\name{US_inaugural_addresses_1993_2025}
\alias{US_inaugural_addresses_1993_2025}
\title{Inaugural speeches from United States presidents from 1993 to 2025}
\format{
A data frame with 9 rows and 3 variables: the president,
the date, and the full text of the inaugural address.
}
\source{
The dataset was gathered from the American Presidency Project
website (https://www.presidency.ucsb.edu/).
}
\usage{
data(US_inaugural_addresses_1993_2025)
}
\description{
A dataset containing nine inaugural addresses since 1993 for
five different United States presidents: Clinton, Bush, Obama, Trump,
and Biden.
}
\keyword{datasets}
