% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/US_News_Conference.R
\docType{data}
\name{US_News_Conferences}
\alias{US_News_Conferences}
\title{US News Conferences Data}
\format{
A data frame with 1210 rows and 5 variables:
\describe{
  \item{title}{title of news conference}
  \item{date}{date of news conference}
  \item{president}{US president}
  \item{source_links}{link to specific news conference}
  \item{text}{text of news conference}
}
}
\source{
\url{https://www.presidency.ucsb.edu/documents/app-categories/presidential/news-conferences}
}
\usage{
US_News_Conferences
}
\description{
A dataset containing the news conferences from US presidents
from 1965 to 2020.
The dataset was gathered from the American Presidency Project
(https://www.presidency.ucsb.edu/).
}
\keyword{datasets}
