% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/US_News_Conferences_1960_1980.R
\docType{data}
\name{US_News_Conferences_1960_1980}
\alias{US_News_Conferences_1960_1980}
\title{US News Conferences Data from 1960 to 1980}
\format{
A data frame with 1210 rows and 5 variables:
\describe{
  \item{president}{US president}
  \item{date}{date of news conference}
  \item{text}{text of news conference}
}
}
\source{
\url{https://www.presidency.ucsb.edu/documents/app-categories/presidential/news-conferences}
}
\usage{
US_News_Conferences_1960_1980
}
\description{
A dataset containing the news conferences from US presidents
from 1960 to 1980.
The dataset was gathered from the American Presidency Project
(https://www.presidency.ucsb.edu/).
}
\keyword{datasets}
