% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_tools.R
\name{annotate_text}
\alias{annotate_text}
\title{Parse text with NLP}
\usage{
annotate_text(v, level = "words")
}
\arguments{
\item{v}{Text vector}

\item{level}{Do you want to parse words or sentences? Words by default.}
}
\description{
Wrapper for `spacyr::spacy_parse` function.
}
\examples{
#annotate_text(US_News_Conferences_1960_1980[1:2, 3])
}
