% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_tools.R
\name{annotate_text}
\alias{annotate_text}
\title{Annotate text with NLP}
\usage{
annotate_text(v, level = "words")
}
\arguments{
\item{v}{Text vector}

\item{level}{Do you want to parse words or sentences? Words by default.}
}
\value{
A data frame with syntax information by words or sentences in text.
}
\description{
This function relies on `{spacyr}` NLP parsing to annotate texts.
}
\examples{
#annotate_text(US_News_Conferences_1960_1980[1:2, 3])
#annotate_text(US_News_Conferences_1960_1980[1:2, 3], level = "sentence")
}
