% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_tools.R
\name{annotate_text}
\alias{annotate_text}
\title{Annotate text with NLP}
\usage{
annotate_text(v, level = "words")
}
\arguments{
\item{v}{Text vector}

\item{level}{At which level would you like to parse the text?
Options include "words" or "sentences".
Defaults to "words".}
}
\value{
A data frame with syntax information by words or sentences in text.
}
\description{
This function relies on `\{spacyr\}` NLP parsing to annotate texts.
}
