% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{context}
\alias{context}
\title{Context to string matches}
\usage{
context(
  string,
  var = NULL,
  context_level = c("sentences", "words", "paragraph"),
  context_size = NULL,
  context_side = c("both", "left", "right"),
  ignore.case = TRUE
)
}
\value{
A list of string matches an their context
}
\description{
A function for getting string matches and the context in which they occur.
}
