% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{context}
\alias{context}
\title{Context to string matches}
\usage{
context(match, textvar, level = c("sentences", "words", "paragraph"), n = 1)
}
\arguments{
\item{match}{Character string to be matched.
For multiple strings, please use "|" as a separator.}

\item{textvar}{Text variable.}

\item{level}{At which text level do you want matches to be returned?
Options are sentences, words, and paragraph.}

\item{n}{Number of sentences or words matched before and after string match.
1 by default.
That is, one word or one sentence before, and after, string match.
For paragraphs, n is always set to one.}
}
\value{
A list of string matches an their context
}
\description{
A function for getting string matches and the context in which they occur.
}
\examples{
context(match = "war|weapons of mass destruction|conflict|NATO|peace",
textvar = US_News_Conferences_1960_1980$text[100],
level = "sentences",
n = 2)
}
