% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{context}
\alias{context}
\title{Context to string matches}
\usage{
context(string, var, level = c("sentences", "words", "paragraph"))
}
\arguments{
\item{string}{Character string to be matched.
For multiple strings, please use "|" as a separator.}

\item{var}{Text variable}

\item{level}{At which text level do you want matches to be returned?
Options are sentences, words, and paragraph.}
}
\value{
A list of string matches an their context
}
\description{
A function for getting string matches and the context in which they occur.
}
\details{
For sentences, the sentence before and the sentence after match are returned.
For words, 3 words before and 3 words after match are returned.
For paragraph, the whole paragraph of the match is returned.
For paragraph to work properly, please make sure paragraph marks are present in text.
}
\examples{
context(string = "war|weapons of mass destruction|conflict|NATO|peace",
var = US_News_Conferences$text[500], level = "sentences")
}
