% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_tools.R
\name{extract_date}
\alias{extract_date}
\title{Extract dates from text}
\usage{
extract_date(v)
}
\arguments{
\item{v}{Text variable/object}
}
\value{
A list of dates
}
\description{
Sometimes dates can be contained in text,
this function extracts those dates from text.
}
\examples{
text <- c("This function was created on the 29 September 2021",
"Today is October 12, 2021")
extract_date(text)
}
