% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_tools.R
\name{extract_date}
\alias{extract_date}
\title{Extract dates from text}
\usage{
extract_date(v, language = "en")
}
\arguments{
\item{v}{Text variable/object}

\item{language}{English is default.
For portuguese, specify "pt".}
}
\value{
A list of dates
}
\description{
Sometimes dates can be contained in text,
this function extracts those dates from text.
}
