% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_tools.R
\name{extract_date}
\alias{extract_date}
\title{Extract dates from text}
\usage{
extract_date(v)
}
\arguments{
\item{v}{Text vector.}
}
\value{
A vector of the dates in text.
}
\description{
Wrapper function for `messydates::as_messydates`.
}
