% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_tools.R
\name{extract_location}
\alias{extract_location}
\title{Extract location from strings}
\usage{
extract_location(v)
}
\arguments{
\item{v}{Text variable/object}
}
\value{
A list of the first locations
}
\description{
Extract location from strings
}
\details{
Works well for Brazilian states and other countries.
Texts must be in English or Portuguese.

If more than one location is found,
returns only the first match.
}
\examples{
extract_location(c("This is the United States", "This is Sao Paulo",
"I was in Rio de Janeiro and Sao Paulo, then back to the United States"))
}
