% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_tools.R
\name{extract_location}
\alias{extract_location}
\title{Extract location from strings}
\usage{
extract_location(v)
}
\arguments{
\item{v}{Text variable/object}
}
\value{
A list of the first locations
}
\description{
Extracts location from strings.
For now, only works for Brasilian states and other
countries/Unions.
}
\examples{
text <- c("This is the US", "This is Sao Paulo")
extract_title(text)
}
