% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_tools.R
\name{extract_match}
\alias{extract_match}
\title{Extract text matches}
\usage{
extract_match(v, match, invert = FALSE, ignore.case = TRUE)
}
\arguments{
\item{v}{Text vector or annotated data frame.}

\item{match}{A regex match for a word(s) or expression.
For multiple words, please use "|" to divide them.}

\item{invert}{Do you want texts without certain matches to be returned?
By default FALSE.}

\item{ignore.case}{Should case be ignored?
By default, TRUE.}
}
\value{
A list the same length as text variable.
}
\description{
Get texts in which certain "matches" occur.
}
\examples{
\donttest{
extract_match(c("This function was created on the 29 September 2021",
"Today is October 12, 2021"), "October")
}
}
