% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment.R
\name{extract_promises}
\alias{extract_promises}
\title{Extract future promises from political discourses}
\usage{
extract_promises(v)
}
\arguments{
\item{v}{Text vector or annotated data frame.}
}
\description{
Extract future promises from political discourses
}
\examples{
\donttest{
extract_promises(US_News_Conferences_1960_1980[1:2,3])
}
}
