% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promises.R
\name{extract_promises}
\alias{extract_promises}
\title{Extract future promises from political discourses}
\usage{
extract_promises(.data)
}
\arguments{
\item{.data}{A (annotated) data frame or text vector.
For data frames, function will search for "text" variable.
For annotated data frames, please declare an annotated data frame
at the sentence level.}
}
\description{
Extract future promises from political discourses
}
\examples{
\donttest{
extract_promises(US_News_Conferences_1960_1980[1:2,3])
}
}
