% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic.R
\name{extract_related_terms}
\alias{extract_related_terms}
\title{Extract terms related to subjects}
\usage{
extract_related_terms(.data, dictionary)
}
\arguments{
\item{.data}{A data frame, promises, or text vector.
For data frames, function will search for "text" variable.
For promises data, function will search for "promises" variable.}

\item{dictionary}{The dictionary of 20 major political topics from the
Comparative Agendas Project is used by default.
Users can also declare a custom dictionary as a vector or a list.
If users declare a vector, each element is treated as a independent topic.
If users declare a list of subjects and related terms, function understands
names as topic and words as terms.}
}
\description{
Extract terms related to subjects
}
\examples{
\donttest{
extract_related_terms(US_News_Conferences_1960_1980[1:5, 3], dictionary = "CAP")
extract_related_terms(US_News_Conferences_1960_1980[1:5, 3],
                      dictionary = c("military", "development"))
extract_related_terms(US_News_Conferences_1960_1980[1:5, 3],
                      dictionary = list("military" = c("military", "gun", "war"),
                                        "development" = c("development", "interest rate", "banks")))
}
}
