% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment.R
\name{extract_related_terms}
\alias{extract_related_terms}
\title{Extract terms related to subjects}
\usage{
extract_related_terms(v, subjects)
}
\arguments{
\item{v}{Text vector or annotated data frame.}

\item{subjects}{Vector containing subjects.}
}
\description{
Extract terms related to subjects
}
\examples{
\donttest{
extract_related_terms(US_News_Conferences_1960_1980[1:2, 3],
                      subjects = extract_subjects(US_News_Conferences_1960_1980[1:2, 3]))
}
}
