% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_tools.R
\name{extract_similarities}
\alias{extract_similarities}
\title{Extract similarities and differences in texts/segments}
\usage{
extract_similarities(v, comparison = "similarities", method)
}
\arguments{
\item{v}{Text vector or annotated data frame.}

\item{comparison}{Would you like to extract similarities or differences
between treaties? Options are "similarities" or "differences".
Defaults to "similarities".}

\item{method}{A method for checking similarities or differences.
For similarities, defaults to "correlation" method.
Other methods from `quanteda.textstats::textstat_simil()`
include "cosine", "jaccard", "ejaccard", "dice", "edice",
"simple matching", and "hamann".
For differences, fedaukts to "euclidean".
Other methods from `quanteda.textstats::textstat_dist()` include
"manhattan", "maximum", "canberra", and "minkowski".}
}
\description{
Extract similarities and differences in texts/segments
}
\examples{
\dontrun{
extract_similarities(US_News_Conferences_1960_1980[1:2,3])
}
}
