% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment.R
\name{extract_subjects}
\alias{extract_subjects}
\title{Extract most frequent subjects from political discourses}
\usage{
extract_subjects(v, n = 20)
}
\arguments{
\item{v}{Text vector.}

\item{n}{Number of subjects}
}
\description{
Extract most frequent subjects from political discourses
}
