% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_speaker.R
\name{get_speakers}
\alias{get_speakers}
\title{Split text by speakers}
\source{
https://stackoverflow.com/questions/41100482/split-speaker-and-dialogue-in-rstudio
The solution described in the forum was used as a reference.
}
\usage{
get_speakers(textvec, splitsign, speakermark)
}
\arguments{
\item{textvec}{text vector}

\item{splitsign}{sign or string used to split text vector.
Often these are line spaces ("/n") but depending on text different
or even string can be chosen (i.e. ",.")}

\item{speakermark}{speaker marker sign.
Often these are colon (":") but depending on text different
or even string can be chosen (i.e. "Thank you" or "I").}
}
\value{
All text aggregated by speaker.
}
\description{
Split text by speakers
}
