% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urgency.R
\name{get_urgency}
\alias{get_urgency}
\title{Urgency Analysis}
\usage{
get_urgency(.data, dictionary, normalize = "tokens")
}
\arguments{
\item{.data}{A data frame, promises, or text vector.
For data frames, function will search for "text" variable.
For promises data, function will search for "promises" variable.}

\item{normalization}{Would you like urgency scores to be normalized?
By default, urgency scores are normalized by "tokens",
the number of words in text observation.
Users can also declare "none", for no normalization.}
}
\value{
A scored data frame.
}
\description{
Urgency Analysis
}
\examples{
\donttest{
get_urgency(US_News_Conferences_1960_1980[1:10,3])
}
}
