% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urgency.R
\name{get_urgency}
\alias{get_urgency}
\title{Urgency Analysis}
\usage{
get_urgency(v, subjects)
}
\arguments{
\item{v}{Text vector or annotated data frame.}

\item{subjects}{List of subjects.}
}
\value{
A scored data frame.
}
\description{
Urgency Analysis
}
\examples{
\donttest{
get_urgency(US_News_Conferences_1960_1980[1:10,3],
            subjects = extract_subjects(US_News_Conferences_1960_1980[1:2, 3]))
}
}
