% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urgency.R
\name{get_urgency}
\alias{get_urgency}
\title{Urgency Analysis}
\usage{
get_urgency(v, n = 20)
}
\arguments{
\item{v}{Text vector.}

\item{n}{Number of subjects.}
}
\value{
A scored data frame.
}
\description{
Urgency Analysis
}
