% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urgency.R
\name{get_urgency}
\alias{get_urgency}
\title{Urgency Analysis}
\usage{
get_urgency(.data, normalize = "tokens")
}
\arguments{
\item{.data}{A data frame, priorities data frame coded using
`select_priorities()`, or text vector.
For data frames, function will search for "text" variable.
For priorities data frame function will search for "priorities" variable.}

\item{normalize}{Would you like urgency scores to be normalized?
By default, urgency scores are normalized by "tokens",
the number of words in text observation.
Users can also declare "none", for no normalization.}
}
\value{
A scored data frame for each dimension of urgency.
}
\description{
Urgency Analysis
}
\examples{
\donttest{
get_urgency(US_News_Conferences_1960_1980[1:10, 3])
get_urgency(US_News_Conferences_1960_1980[1:10,])
}
}
