% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urgency.R
\name{get_urgency_rank}
\alias{get_urgency_rank}
\title{Rank urgent topics}
\usage{
get_urgency_rank(v, subjects)
}
\arguments{
\item{subjects}{List of subjects.}

\item{.data}{A data frame, promises, or text vector.
For data frames, function will search for "text" variable.
For promises data, function will search for "promises" variable.}
}
\description{
Rank urgent topics
}
\examples{
\donttest{
get_urgency_rank(US_News_Conferences_1960_1980[1:10,3])
}
}
