% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priorities.R
\name{select_priorities}
\alias{select_priorities}
\title{Select future priorities from political discourses}
\usage{
select_priorities(.data, na.rm = TRUE)
}
\arguments{
\item{.data}{A (annotated) data frame or text vector.
For data frames, function will search for "text" variable.
For annotated data frames, please declare an annotated data frame
at the sentence level.}

\item{na.rm}{Would you like political statements that do not contain a
political action to be removed?
By default, TRUE.}
}
\value{
A data frame with syntax information by sentences and
a variable identifying which of these sentences are priorities.
}
\description{
Political priorities are statements in which actors express their
intent or commitment to take political action in the future.
}
