% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urgency.R
\name{sim_urgency}
\alias{sim_urgency}
\title{Simulating urgency}
\usage{
sim_urgency(urgency, commitment, intensity, timing, frequency, pronoun = "We")
}
\description{
Simulating urgency
}
\examples{
sim_urgency()
}
