% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment.R
\name{split_text}
\alias{split_text}
\title{Split texts}
\usage{
split_text(text, splitsign = "\\\\.")
}
\arguments{
\item{text}{text variable}

\item{splitsign}{Where do you want to split?
By default sentences (".").
This can also be words, signals or other markers you want.
For special characters, please use escape sign before (i.e. "\\").}
}
\value{
A splitted list for each row
}
\description{
Split texts into structured lists of lists according to a split sign.
}
\examples{
text <- "This is the first sentence. This is the second sentence."
split_text(text)
}
