% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/US_News_Conferences_1960_1980.R
\docType{data}
\name{US_News_Conferences_1960_1980}
\alias{US_News_Conferences_1960_1980}
\title{US News Conferences Data from 1960 to 1980}
\format{
A data frame with 353 rows and 3 variables: the president,
the date, and the full text.
}
\usage{
data(US_News_Conferences_1960_1980)
}
\description{
A dataset containing the news conferences from US presidents
from 1960 to 1980.
The dataset was gathered from the American Presidency Project website.
}
\keyword{datasets}
